Render all levels in QTree visualization
Since we have more optimized pixel map for quad tree,
and the new version of druid seems to be faster, it's worth rendering
everything. After all, it is debug visualization, so the reflection of
reality is prioritized.

Technically, it should be possible to optimize the rendering by
utilizing quad tree's spatial locality, but I don't bother doint it
right now.