Add "Reset all" button to reset all files to initial state
In order to recover from a situation that the user deleted or messed up
files to the point that the game is not functional.
If you run the application locally with `cargo run`, the virtual file
system will write to actual files, so the recovery would be done by
`git reset --h`.