Find path to the closest fog, gradually exploring unknown
It is faster and more efficient to explore than picking closest fog
pixel by Euclidean distance. It uses quadtree to speed up the search,
and also find any fog pixel that has encountered.

Note that fog pixels are defined by pixels, not by quadtree nodes, so
searching may not find the exactly closest fog pixel, but if the area is
open enough, the difference shouldn't matter.